refactored test for simple postgres connection
